Move out of frames layer into association package
