+++ src/main/kotlin/solve/scene/view/GridProvider.kt
@@ -2,6 +2,7 @@ package solve.scene.view
 
 import javafx.beans.property.DoubleProperty
 import solve.scene.model.VisualizationFrame
+import solve.scene.view.association.OutOfFramesLayer
 
 interface GridProvider {
     fun createGrid(
